+++ CHANGELOG.md
@@ -35,6 +35,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Removed unused beta channel
 * Dynamically create instance specific service account, role, and role binding
 * OpenShift deployment now allows the operator to run for all namespaces
+* Now validates if `pvc.Volumemode` set correctly to `Filesystem`
 
 ## Fixed